# content: make the content cache a broker module