--- a/etc/rc1
+++ b/etc/rc1
@@ -22,6 +22,7 @@ if test $RANK -gt 0; then
     flux config reload
 fi
 
+modload all content
 modload all barrier
 if test "$(flux config get --default=false systemd.enable)" = "true"; then
     modload all sdbus