--- a/src/broker/Makefile.am
+++ b/src/broker/Makefile.am
@@ -40,12 +40,6 @@ libbroker_la_SOURCES = \
 	attr.c \
 	log.h \
 	log.c \
-	content-cache.h \
-	content-cache.c \
-	content-checkpoint.h \
-	content-checkpoint.c \
-	content-mmap.h \
-	content-mmap.c \
 	runat.h \
 	runat.c \
 	state_machine.h \