--- a/src/modules/content/cache.h
+++ b/src/modules/content/cache.h
@@ -8,18 +8,16 @@
  * SPDX-License-Identifier: LGPL-3.0
 \************************************************************/
 
-#ifndef HAVE_BROKER_CONTENT_CACHE_H
-#define HAVE_BROKER_CONTENT_CACHE_H 1
-
-#include "attr.h"
+#ifndef _CONTENT_CACHE_H
+#define _CONTENT_CACHE_H 1
 
 struct content_cache;
 
-struct content_cache *content_cache_create (flux_t *h, attr_t *attr);
+struct content_cache *content_cache_create (flux_t *h);
 void content_cache_destroy (struct content_cache *cache);
 bool content_cache_backing_loaded (struct content_cache *cache);
 
-#endif /* !HAVE_BROKER_CONTENT_CACHE_H */
+#endif /* !_CONTENT_CACHE_H */
 
 /*
  * vi:tabstop=4 shiftwidth=4 expandtab