Problem: the content service embedded in the broker is resource-intensive under load, potentially impacting broker performance.  Furthermore, it is a complex, optional service rather than an integral broker component.
    
Move the content service to a broker module where it is less tightly coupled to the broker and has a private reactor thread.

I didn't see any obvious impact on performance when running the throughput test.

**This branch**
```
$ src/test/throughput.py -n 10000
number of jobs: 10000
submit time:    2.964 s (3373.6 job/s)
script runtime: 45.773s
job runtime:    45.616s
throughput:     219.2 job/s (script: 218.5 job/s)

$ src/test/throughput.py -n 10000 -x
number of jobs: 10000
submit time:    3.216 s (3109.4 job/s)
script runtime: 449.663s
job runtime:    103.558s
throughput:     96.6 job/s (script:  22.2 job/s)
```

**Master (v0.54.0-32-g9d9216681)**
```
$ src/test/throughput.py -n 10000
number of jobs: 10000
submit time:    3.008 s (3324.5 job/s)
script runtime: 46.227s
job runtime:    46.054s
throughput:     217.1 job/s (script: 216.3 job/s)

$ src/test/throughput.py -n 10000 -x
number of jobs: 10000
submit time:    2.850 s (3509.2 job/s)
script runtime: 460.330s
job runtime:    93.438s
throughput:     107.0 job/s (script:  21.7 job/s)
```