@garlick 2023-09-10
The accounting test won't pass until flux-framework/flux-accounting#383 is merged.