@garlick 2023-09-12
Now that flux-framework/flux-accounting#383 is merged I'll set MWP here.  Thanks!